+++ optimize.c
@@ -11,7 +11,7 @@
 #include "strings.h"
 
 // optimize numbers in the range (-99.5,99.5) to 2 significant digits
-bool optimize_float_2_sig_dig (const char *snip, unsigned len, double cap_value_at /* 0 if no cap */,
+bool optimize_float_2_sig_dig (const char *snip, unsigned len, float cap_value_at /* 0 if no cap */,
                                char *optimized_snip, unsigned *optimized_snip_len)
 {
     if (!IS_DIGIT(snip[0]) && snip[0] != '.' && snip[0] != '-') return false; // not a number
@@ -20,7 +20,7 @@ bool optimize_float_2_sig_dig (const char *snip, unsigned len, double cap_value_
 
     // temporarily nul-terminate string and get number
     SAFE_NUL (&snip[len]);
-    double fp = atof (snip);
+    float fp = atof (snip);
     SAFE_RESTORE;
 
     // cap value if requested
@@ -36,8 +36,8 @@ bool optimize_float_2_sig_dig (const char *snip, unsigned len, double cap_value_
     #define NUM_EXPS 8
     #define MAX_NUM_LEN (1 /* hyphen */ + (NUM_EXPS-1) /* prefix */ + 2 /* digits */)
 
-    static const double exps[NUM_EXPS]    = { 10, 1, 0.1, 0.01, 0.001, 0.0001, 0.00001, 0.000001 };
-    static const double mult_by[NUM_EXPS] = { 1, 10, 100, 1000, 10000, 100000, 1000000, 10000000 };
+    static const float exps[NUM_EXPS]    = { 10, 1, 0.1, 0.01, 0.001, 0.0001, 0.00001, 0.000001 };
+    static const float mult_by[NUM_EXPS] = { 1, 10, 100, 1000, 10000, 100000, 1000000, 10000000 };
     static const char *prefix = "0.0000000000000000000";
     unsigned e=0; for (; e < NUM_EXPS; e++)
         if (fp >= exps[e]) {
