+++ progress.h
@@ -14,8 +14,8 @@ extern void progress_update (char **prefix, uint64_t sofar, uint64_t total, bool
 extern void progress_update_status (char **prefix, const char *status);
 extern void progress_finalize_component (const char *status);
 extern void progress_finalize_component_time (const char *status, Digest md5);
-extern void progress_finalize_component_time_ratio (const char *me, double ratio, Digest md5);
-extern void progress_finalize_component_time_ratio_better (const char *me, double ratio, const char *better_than, double ratio_than, Digest md5);
+extern void progress_finalize_component_time_ratio (const char *me, float ratio, Digest md5);
+extern void progress_finalize_component_time_ratio_better (const char *me, float ratio, const char *better_than, float ratio_than, Digest md5);
 extern void progress_concatenated_md5 (const char *me, Digest md5);
 
 #endif
